requirements.txt
request
lxml
pyautogui
selenium
pyinquirer
# promt_toolkit
pyfiglet